LPS-202141 move class to follow same pattern on taglibs
