+++ modules/apps/sharing/sharing-taglib/src/main/java/com/liferay/sharing/taglib/internal/permission/util/SharingPermissionUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.sharing.taglib.internal.servlet;
+package com.liferay.sharing.taglib.internal.permission.util;
 
 import com.liferay.portal.kernel.module.service.Snapshot;
 import com.liferay.sharing.security.permission.SharingPermission;
