+++ modules/apps/sharing/sharing-taglib/src/main/java/com/liferay/sharing/taglib/servlet/taglib/SharingButtonTag.java
@@ -13,9 +13,9 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.sharing.display.context.util.SharingJavaScriptFactory;
 import com.liferay.sharing.security.permission.SharingPermission;
+import com.liferay.sharing.taglib.internal.permission.util.SharingPermissionUtil;
 import com.liferay.sharing.taglib.internal.servlet.ServletContextUtil;
 import com.liferay.sharing.taglib.internal.servlet.SharingJavaScriptFactoryUtil;
-import com.liferay.sharing.taglib.internal.servlet.SharingPermissionUtil;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.jsp.PageContext;
